R: Moved thunks to separate classes
